Moved UpsetPlot to a new python script specialized in graphs
